dom/base/test/test_bug749367.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=749367
-->

<head>
  <title>Test for Bug 749367</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>

<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=749367">Mozilla Bug 749367</a>
<div id="content" style="display: none"></div>
  <pre id="test">

  <script type="text/template">
  ok(false, "Shouldn't execute");
  </script>

  <script type="text/javascript">
  ok(true, "Should execute");
  </script>

  </pre>
</body>

</html>
